refactor: RandomLineGenerator 클래스에 존재하는 validator도 validator클래스로 이동
