+++ src/main/java/kr/codesquad/util/RandomLineGenerator.java
@@ -7,13 +7,17 @@ import java.util.stream.IntStream;
 
 public class RandomLineGenerator {
 
-    private static final String ERROR_MESSAGE_FOR_CONSECUTIVE_POINTS = "연속적인 true 존재";
+    private static Validator validator;
+
+    public RandomLineGenerator(){
+        this.validator = new Validator();
+    }
 
     public Boolean buildLadderLine(int lineWidth,ArrayList<Boolean> points) {
         IntStream.range(0, lineWidth)
                 .mapToObj(i -> generateRungStatus(points))
                 .forEach(points::add);
-        return validatePoints(points);
+        return validator.validatePoints(points);
     }
 
     private boolean generateRungStatus(List<Boolean> points) {
@@ -22,24 +26,4 @@ public class RandomLineGenerator {
         }
         return new Random().nextBoolean() && !points.get(points.size() - 1);
     }
-
-    private boolean validatePoints(ArrayList<Boolean> points) {
-        try {
-            validatePointsAndThrowException(points);
-            return true;
-        }catch (IllegalStateException e){
-            System.out.println(e);
-        }
-        return false;
-    }
-
-    private void validatePointsAndThrowException(ArrayList<Boolean> points){
-        IntStream.range(0, points.size()-1)
-                .filter(i ->points.get(i) && points.get(i + 1))
-                .findFirst()
-                .ifPresent(i -> {
-                    throw new IllegalStateException(ERROR_MESSAGE_FOR_CONSECUTIVE_POINTS);
-                });
-    }
-
 }
